1) Create unit test cases for poster-related services and utilities 2) Rename a few of classes

+++ work/src/org/kuali/kfs/module/ld/dataaccess/impl/LaborLedgerEntryDaoOjb.java
@@ -61,7 +61,7 @@ public class LaborLedgerEntryDaoOjb extends PersistenceBrokerDaoSupport implemen
         if (iterator.hasNext()) {
             Object[] data = (Object[]) iterator.next();
             if (data[0] != null) {
-                maxSequenceNumber = (Integer)data[0];
+                maxSequenceNumber = ((BigDecimal)data[0]).intValue();
             }
         }
         return maxSequenceNumber;
